Adds `RemoteFeedItem` & refactor module - Renames `FeedItemsMapper` to `RemoteFeedMapper` - Decouples mapper from the `FeedItem` data model & feature interface
